+++ codex-rs/core/src/codex_tests_guardian.rs
@@ -1,0 +1,153 @@
+use super::*;
+use crate::exec::ExecParams;
+use crate::features::Feature;
+use crate::protocol::AskForApproval;
+use crate::sandboxing::SandboxPermissions;
+use crate::turn_diff_tracker::TurnDiffTracker;
+use codex_protocol::models::FunctionCallOutputBody;
+use codex_protocol::models::NetworkPermissions;
+use codex_protocol::models::PermissionProfile;
+use pretty_assertions::assert_eq;
+use serde::Deserialize;
+use std::collections::HashMap;
+use std::sync::Arc;
+
+#[tokio::test]
+async fn guardian_allows_shell_additional_permissions_requests_past_policy_validation() {
+    let (mut session, mut turn_context_raw) = make_session_and_context().await;
+    turn_context_raw
+        .approval_policy
+        .set(AskForApproval::Guardian)
+        .expect("test setup should allow updating approval policy");
+    session
+        .features
+        .enable(Feature::RequestPermissions)
+        .expect("test setup should allow enabling request permissions");
+    turn_context_raw
+        .sandbox_policy
+        .set(SandboxPolicy::DangerFullAccess)
+        .expect("test setup should allow updating sandbox policy");
+    let session = Arc::new(session);
+    let turn_context = Arc::new(turn_context_raw);
+
+    let params = ExecParams {
+        command: if cfg!(windows) {
+            vec![
+                "cmd.exe".to_string(),
+                "/C".to_string(),
+                "echo hi".to_string(),
+            ]
+        } else {
+            vec![
+                "/bin/sh".to_string(),
+                "-c".to_string(),
+                "echo hi".to_string(),
+            ]
+        },
+        cwd: turn_context.cwd.clone(),
+        expiration: 1000.into(),
+        env: HashMap::new(),
+        network: None,
+        sandbox_permissions: SandboxPermissions::WithAdditionalPermissions,
+        windows_sandbox_level: turn_context.windows_sandbox_level,
+        justification: Some("test".to_string()),
+        arg0: None,
+    };
+
+    let handler = ShellHandler;
+    let resp = handler
+        .handle(ToolInvocation {
+            session: Arc::clone(&session),
+            turn: Arc::clone(&turn_context),
+            tracker: Arc::new(tokio::sync::Mutex::new(TurnDiffTracker::new())),
+            call_id: "test-call".to_string(),
+            tool_name: "shell".to_string(),
+            payload: ToolPayload::Function {
+                arguments: serde_json::json!({
+                    "command": params.command.clone(),
+                    "workdir": Some(turn_context.cwd.to_string_lossy().to_string()),
+                    "timeout_ms": params.expiration.timeout_ms(),
+                    "sandbox_permissions": params.sandbox_permissions,
+                    "additional_permissions": PermissionProfile {
+                        network: Some(NetworkPermissions {
+                            enabled: Some(true),
+                        }),
+                        file_system: None,
+                        macos: None,
+                    },
+                    "justification": params.justification.clone(),
+                })
+                .to_string(),
+            },
+        })
+        .await;
+
+    let output = match resp.expect("expected Ok result") {
+        ToolOutput::Function {
+            body: FunctionCallOutputBody::Text(content),
+            ..
+        } => content,
+        _ => panic!("unexpected tool output"),
+    };
+
+    #[derive(Deserialize, PartialEq, Eq, Debug)]
+    struct ResponseExecMetadata {
+        exit_code: i32,
+    }
+
+    #[derive(Deserialize)]
+    struct ResponseExecOutput {
+        output: String,
+        metadata: ResponseExecMetadata,
+    }
+
+    let exec_output: ResponseExecOutput =
+        serde_json::from_str(&output).expect("valid exec output json");
+
+    assert_eq!(exec_output.metadata, ResponseExecMetadata { exit_code: 0 });
+    assert!(exec_output.output.contains("hi"));
+}
+
+#[tokio::test]
+async fn guardian_allows_unified_exec_additional_permissions_requests_past_policy_validation() {
+    let (mut session, mut turn_context_raw) = make_session_and_context().await;
+    turn_context_raw
+        .approval_policy
+        .set(AskForApproval::Guardian)
+        .expect("test setup should allow updating approval policy");
+    session
+        .features
+        .enable(Feature::RequestPermissions)
+        .expect("test setup should allow enabling request permissions");
+    let session = Arc::new(session);
+    let turn_context = Arc::new(turn_context_raw);
+    let tracker = Arc::new(tokio::sync::Mutex::new(TurnDiffTracker::new()));
+
+    let handler = UnifiedExecHandler;
+    let resp = handler
+        .handle(ToolInvocation {
+            session: Arc::clone(&session),
+            turn: Arc::clone(&turn_context),
+            tracker: Arc::clone(&tracker),
+            call_id: "exec-call".to_string(),
+            tool_name: "exec_command".to_string(),
+            payload: ToolPayload::Function {
+                arguments: serde_json::json!({
+                    "cmd": "echo hi",
+                    "sandbox_permissions": SandboxPermissions::WithAdditionalPermissions,
+                    "justification": "need additional sandbox permissions",
+                })
+                .to_string(),
+            },
+        })
+        .await;
+
+    let Err(FunctionCallError::RespondToModel(output)) = resp else {
+        panic!("expected validation error result");
+    };
+
+    assert_eq!(
+        output,
+        "missing `additional_permissions`; provide at least one of `network`, `file_system`, or `macos` when using `with_additional_permissions`"
+    );
+}
